Trait regalloc::PrettyPrintSized

pub trait PrettyPrintSized: PrettyPrint {
    fn show_rru_sized(
        &self, 
        maybe_reg_universe: Option<&RealRegUniverse>, 
        _size: u8
    ) -> String { ... }
}

Same as PrettyPrint, but can also take a size hint into account to specialize the displayed string.

Provided methods

fn show_rru_sized(
&self,
maybe_reg_universe: Option<&RealRegUniverse>,
_size: u8
) -> String

The same as |show_rru|, but with an optional hint giving a size in bytes. Its interpretation is object-dependent, and it is intended to pass around enough information to facilitate printing sub-parts of real registers correctly. Objects may ignore size hints that are irrelevant to them.

The default implementation ignores the size hint.